data-pipeline/requirements.txt
azure-core==1.32.0
azure-storage-blob==12.23.1
boto3==1.35.53
botocore==1.35.53
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
cryptography==43.0.3
idna==3.10
isodate==0.7.2
jmespath==1.0.1
numpy==2.1.2
pandas==2.2.3
py4j==0.10.9.7
pycparser==2.22
pyspark==3.5.3
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
requests==2.32.3
s3transfer==0.10.3
six==1.16.0
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
